--- frontend/authentication/src/App.css
@@ -1,119 +1,0 @@
-
-.App-header {
-  background-color: #282c34;
-  min-height: 100vh;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  color: white;
-}
-
-.App-link {
-  color: #61dafb;
-}
-
-.loginP , .forgotP , .registerP{
-    font-size: 1.8rem;
-    background-color:rgb(119, 61, 212);
-    border: none;
-}
-
-.btn:hover {
-    cursor: pointer;
-    background-color: rgb(110, 100, 199);
-}
-
-.main-container {
-  width: 100%;
-  height: 35rem;
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  overflow: hidden;
-  margin-top: 50px;
-  border-radius: 5px;
-  box-shadow: 30px 3px 60px rgba(8, 16, 19, 0.3);
-  background-color: rgb(211, 203, 203);
-}
-
-@keyframes slide-down {
-  0% {
-    transform: translateY(-100%);
-  }
-  100% {
-    transform: translateY(0%);
-  }
-}
-
-.form-container {
-  width: 100%;
-  height: 100%;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  animation: slide-up 0.5s ease;
-}
-
-@keyframes slide-up {
-  0% {
-    transform: translateY(100%);
-  }
-  100% {
-    transform: translateY(0);
-  }
-}
-
-form a {
-
-  font-size: 1.5rem;
-  text-decoration: none;
-}
-
-form span {
-  font-size: 1.6rem;
-  display: block;
-}
-
-form a:hover {
-    color:red;
-}
-
-.register-a:hover {
-    color: blue;
-}
-.register {
-  display: none;
-}
-
-.forgot {
-  display: none;
-  position: relative;
-}
-
-.iconclose {
-  position: absolute;
-  top: 1.3rem;
-  left: 1.5rem;
-  border: 1px solid red;
-  border-radius: 50%;
-  height: 3rem;
-  width: 3rem;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  cursor: pointer;
-}
-.fa-times {
-    color: red;
-    font-size: 1.5rem;
-}
-@media screen and (max-width: 720px){
-    .img-container{
-        display: none;
-
-    }
-    .form-container {
-        width: 100%;
-    }
-}
